api_django/requirements.txt
affine==2.3.1
amqp==5.1.1
asgiref==3.5.0
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
earthengine-api==0.1.319
future==0.18.2
geojson==2.5.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
jdatetime==4.1.0
kombu==5.2.4
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
rasterio==1.3.0.post1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
Shapely==1.8.2
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.2
tifffile==2022.8.12
tzdata==2022.2
uritemplate==3.0.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
